Tweak new test locations
